Lab05/lab5_4/kafka_quotes/pyproject.toml
[tool.poetry]
name = "kafka-quotes"
version = "0.1.0"
description = ""
authors = ["Tomás Santos <t.santos.f@ua.pt>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
kafka-python = "^2.0.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
